backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
folium==0.14.0
asyncpg==0.29.0
python-dotenv==1.0.0
